feat: Calculate the sentiment of the text in the overview column
